Extract all test utils
